requirements.txt
psutil>=5.9.0
pycapnp>=1.3.0
numpy>=1.24.0
numba>=0.59.0
